# src/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Enterprise Application API
  version: 1.0.0
  description: This is the OpenAPI specification for the enterprise application.
servers:
  - url: https://api.example.com/v1
    description: Production server
paths:
  /api/users:
    get:
      summary: Retrieve user list
      responses:
        '200':
          description: A list of users.
    post:
      summary: Create new user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                email:
                  type: string
      responses:
        '201':
          description: User created successfully.
  /api/auth/login:
    post:
      summary: User login
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: User logged in successfully.
...